+++ ts/image-picker/image-picker.translates.ts
@@ -3,9 +3,9 @@ module lui.imagepicker {
 	/* tslint:disable */
 	angular.module("lui.crop").config(["$translateProvider", function ($translateProvider) {
 		$translateProvider.translations("en", {
-			"LUIIMGPICKER_UPLOAD_IMAGE": "change picture",
-			"LUIIMGPICKER_MODIFY_IMAGE": "modify picture",
-			"LUIIMGPICKER_DELETE_IMAGE": "delete picture",
+			"LUIIMGPICKER_UPLOAD_IMAGE": "Change image",
+			"LUIIMGPICKER_MODIFY_IMAGE": "Upload an image",
+			"LUIIMGPICKER_DELETE_IMAGE": "Remove image",
 			"LUIIMGCROPPER_CROP": "Crop",
 			"LUIIMGCROPPER_DO_NOT_CROP": "Do not crop",
 		});
@@ -16,9 +16,9 @@ module lui.imagepicker {
 
 		});
 		$translateProvider.translations("fr", {
-			"LUIIMGPICKER_UPLOAD_IMAGE": "changer l'image",
-			"LUIIMGPICKER_MODIFY_IMAGE": "modifier l'image",
-			"LUIIMGPICKER_DELETE_IMAGE": "supprimer l'image",
+			"LUIIMGPICKER_UPLOAD_IMAGE": "Changer l'image",
+			"LUIIMGPICKER_MODIFY_IMAGE": "Télécharger une nouvelle image",
+			"LUIIMGPICKER_DELETE_IMAGE": "Supprimer l'image",
 			"LUIIMGCROPPER_CROP": "Recadrer",
 			"LUIIMGCROPPER_DO_NOT_CROP": "Ne pas recadrer",
 		});